Skip multi-host setup for nova network setup
There is a workaround which always set multi-hosts for Nova
network. It creates a br_flat interface which is connected
to vmnet and DHCP server listens on this interface. But
XenServer has eth3 connected to vmnet. We should use eth3
as the flat interface otherwise the devstack VM has no access
to the guest instances. So we shouldn't apply that workaround
for XenServer. And we currently don't use multi-hosts for
XenServer CI. So let's skipt it.